requirements.txt
langchain
pinecone-client
python-docx
PyPDF2
streamlit
streamlit_chat
unstructured
openai
tiktoken
pypdf
langchain[docarray]
Flask
Requests
duckduckgo-search
openpyxl
yfinance
flask-cors
nltk